requirements.txt
filelock==3.15.4
fsspec==2024.6.1
intel-openmp==2021.4.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.19.5
pillow==10.3.0
sympy==1.12.1
tbb==2021.13.0
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.12.2